control-plane/detector/requirements.txt
paho-mqtt>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
scikit-learn>=1.4,<2.0
numpy>=1.26,<3.0
